rem: 기존 MainPage 제거 [CAV-102]
