+++ src/pages/MainPage.vue
@@ -7,144 +7,31 @@ import { today, getLocalTimeZone } from '@internationalized/date';
 import { Calendar } from '@/components/ui/calendar';
 import Card from '@/components/Card.vue';
 import UserCard from '@/components/UserCard.vue';
-import PieChart from '@/components/PieChart.vue';
 import LineChart from '@/components/LineChart.vue';
 import DashboardContainer from '@/components/CardsContainer.vue';
-import { useTransactionStore } from '@/stores/transaction.js';
 import { useUserStore } from '@/stores/user.js';
 import ToggleButton from '@/components/ToggleButton.vue';
 import BarChart from '@/components/BarChart.vue';
+import { useBarChartStore } from '@/stores/barChart';
 import TimeLine from '@/components/tradeList/TimeLine.vue';
-const transaction = useTransactionStore();
-const userStore = useUserStore();
+
+import { useDurationStore } from '@/stores/duration';
 // 상태 관리
-const date = ref(today(getLocalTimeZone()));
-const duration = ref('month'); // 'month', 'week', 'day'
-
-// 전체 송금 데이터 (반응형 상태)
-const monthlyTrans = computed(() => transaction.monthlyTrans);
-const durationTrans = computed(() => transaction.durationTransactions);
-
-// 사용자 정보
+const userStore = useUserStore(); // 유저 정보 관리
 const user = computed(() => userStore.user);
+const durationStore = useDurationStore(); // 수입/지출 내역 기간 관리
+const barChartStore = useBarChartStore(); // barChart 데이터 관리
 
-// 총 금액 계산
-const totalAmount = computed(() => {
-  if (!monthlyTrans.value || monthlyTrans.value.length === 0) return 0;
-  return monthlyTrans.value.reduce((sum, t) => sum + t.amount, 0);
+const date = computed({
+  get: () => durationStore.date,
+  set: (val) => (durationStore.date = val),
 });
 
 // 날짜별 수입/지출 횟수
-const dateTransactionNumber = computed(() => transaction.dateTransactionNumber);
+const dateTransactionNumber = computed(() => durationStore.dateTransactionNumber);
+const barIncome = computed(() => barChartStore.barIncome);
+const barExpense = computed(() => barChartStore.barExpense);
 
-const barExpense = computed(() => {
-  const sumByCategory = durationTrans.value?.expense.reduce((acc, t) => {
-    const category = t.category || '미분류';
-    acc[category] = (acc[category] || 0) + t.amount;
-    return acc;
-  }, {});
-
-  const palette = [
-    '#EF4444', // 짙은 빨간색 (가장 큰 금액)
-    '#F87171', // 진한 레드
-    '#FCA5A5', // 부드러운 코랄 레드
-    '#FECACA', // 밝은 핑크 레드
-    '#FEE2E2', // 연한 파스텔 레드
-    '#FEF2F2', // 아주 연하고 투명한 레드
-  ];
-
-  return Object.entries(sumByCategory)
-    .sort((a, b) => b[1] - a[1])
-    .map(([category, sum], index) => ({
-      type: category,
-      value: sum,
-      color: palette[index % palette.length],
-    }));
-});
-
-const barIncome = computed(() => {
-  const sumByCategory = durationTrans.value?.income.reduce((acc, t) => {
-    const category = t.category || '미분류';
-    acc[category] = (acc[category] || 0) + t.amount;
-    return acc;
-  }, {});
-
-  const palette = [
-    '#10B981', // 짙은 에메랄드 그린 (가장 큰 금액)
-    '#34D399', // 진한 민트 그린
-    '#6EE7B7', // 부드러운 녹색
-    '#A7F3D0', // 밝은 파스텔 그린
-    '#D1FAE5', // 연한 민트
-    '#ECFDF5', // 아주 연한 화이트 그린
-  ];
-
-  return Object.entries(sumByCategory)
-    .sort((a, b) => b[1] - a[1])
-    .map(([category, sum], index) => ({
-      type: category,
-      value: sum,
-      color: palette[index % palette.length],
-    }));
-});
-
-// 날짜 포맷팅 헬퍼 함수 (YYYY-MM-DDTHH:mm:ss 형식)
-const formatDate = (dateObj, isEnd = false) => {
-  const y = dateObj.getFullYear();
-  const m = String(dateObj.getMonth() + 1).padStart(2, '0');
-  const d = String(dateObj.getDate()).padStart(2, '0');
-  return isEnd ? `${y}-${m}-${d}T23:59:59` : `${y}-${m}-${d}T00:00:00`;
-};
-
-// 선택된 날짜와 duration에 따라 startDate와 endDate 계산
-const dateRange = computed(() => {
-  // @internationalized/date 객체를 JS 기본 Date 객체로 변환
-  const current = date.value.toDate(getLocalTimeZone());
-  const year = current.getFullYear();
-  const month = current.getMonth();
-  const day = current.getDate();
-
-  let start, end;
-
-  if (duration.value === 'month') {
-    start = new Date(year, month, 1); // 월의 첫째 날
-    end = new Date(year, month + 1, 0); // 월의 마지막 날
-  } else if (duration.value === 'week') {
-    const dayOfWeek = current.getDay(); // 0(일) ~ 6(토)
-    const diffToStart = day - dayOfWeek; // 일요일을 주의 시작으로 기준
-    start = new Date(year, month, diffToStart);
-    end = new Date(year, month, diffToStart + 6); // 토요일
-  } else {
-    // 'day'
-    start = new Date(year, month, day);
-    end = new Date(year, month, day);
-  }
-
-  return {
-    startDate: formatDate(start, false),
-    endDate: formatDate(end, true),
-  };
-});
-
-// dateRange 값이 바뀔 때마다(또는 처음 렌더링될 때) API 호출
-watch(
-  () => [dateRange.value, user.value?.userid],
-  ([newRange, userId]) => {
-    // 💡 유저 정보가 아직 로딩 전이라면 API를 호출하지 않고 기다립니다.
-    if (!userId) return;
-
-    // 유저 아이디가 확인되면 그제서야 안전하게 API를 호출합니다.
-    transaction.getUserAllTransaction(userId, newRange.startDate, newRange.endDate);
-    transaction.getDurationTransaction(userId, newRange.startDate, newRange.endDate);
-  },
-  { immediate: true, deep: true }
-);
-
-const handleDurationChange = (selectedValue) => {
-  // console.log('선택된 기간:', selectedValue); // 'month', 'week', 'day'
-  duration.value = selectedValue;
-};
-
-const myValue = ref(75);
 const myWeeklyData = ref([
   { label: 'Mon', value: 30 },
   { label: 'Tue', value: 50 },
@@ -186,7 +73,7 @@ const resetLayout = () => {
       </div> -->
       <ToggleButton
         @reset="resetLayout"
-        @duration="handleDurationChange"
+        @duration="durationStore.handleDurationChange"
       ></ToggleButton>
 
       <!-- 드래그 영역 -->
